+++ src/config/remove_app.js
@@ -4,8 +4,8 @@ export function remove_app(app_instance) {
     app_instance.$off()
     app_instance.$el.remove()
     app_instance = null
-    console.log('app_instance destroyed')
+    console.info('[switching app] app_instance destroyed')
   } else {
-    console.log('app_instance DOES NOT exists - no need to remove',)
+    console.info('[switching app] app_instance DOES NOT exists - no need to remove',)
   }
 }
