+++ src/shell_app/models/users/controller.js
@@ -1,16 +1,19 @@
+import {get, merge} from 'lodash'
+
 import Remote from './remote'
+import {store} from 'shell_app/launch_shell_app'
 
-async function get_permissions({user, instance_config}) {
-  const res = await Remote.permissions({user_id: user.id})
-
-  console.log('fake get permissions for', user._id, 'on', instance_config.instance.slug)
-  const applets = ['irs_monitor', 'irs_plan', 'irs_record_point', 'irs_tasker', 'record2', 'debug', 'bod-client']
-  return Promise.resolve([
-    ...applets.map(a => 'read:' + a),
-    ...applets.map(a => 'write:' + a),
-  ])
+async function add_permissions({user, instance_id}) {
+  const res = await Remote.permissions({user_id: user.id, instance_id})
+  const permissions = get(res, 'data.permissions')
+  if (permissions) {
+    const user_with_permissions = merge(user, {permissions})
+    store.commit('set_user', user_with_permissions)
+    return user_with_permissions
+  }
+  return false
 }
 
 export default {
-  get_permissions,
+  add_permissions,
 }
