+++ src/shell_app/models/users/remote.js
@@ -2,13 +2,9 @@ import {shell_axios} from '../../lib/shell_request_handler'
 
 const prefix = 'users'
 
-async function permissions({user_id}) {
+async function permissions({user_id, instance_id}) {
   const request = {
-    method: 'get',
-    url_suffix: `/${prefix}/permissions/${user_id}`,
-    headers: {
-      Authorization: 'some_token'
-    }
+    url: `/${prefix}/${user_id}/permissions/${instance_id}`,
   }
   return shell_axios(request)
 }
