+++ src/shell_app/pages/ToolbarControls.vue
